refactor: PotForm 헤더 컴포넌트화
